+++ ocdsmerge/__init__.py
@@ -1,1 +1,1 @@
-from .merge import merge, merge_versioned, get_merge_rules, WARN_ON_COLLISION, RAISE_ON_COLLISION, IGNORE_ON_COLLISION  # noqa
+from .merge import merge, merge_versioned, get_merge_rules, WARN, RAISE, IGNORE  # noqa